% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calculer la distribution des ages}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{Un data.frame contenant les données des élus, incluant la colonne \code{Date_de_naissance}
qui spécifie la date de naissance des élus.}
}
\value{
Un vecteur contenant les quantiles des âges (min, 25e percentile, médiane, 75e percentile, max).
}
\description{
Cette fonction calcule la distribution des âges des élus à partir de leur date de naissance.
Elle transforme d'abord la colonne \code{Date_de_naissance} en format Date, puis calcule l'âge
de chaque élu en années, et retourne les quantiles (minimum, 25e percentile, médiane, 75e percentile, maximum) des âges.
}
\examples{
df <- data.frame(Date_de_naissance = c("01/01/1980", "15/02/1990", "20/03/1975"))
calcul_distribution_age(df)

}
