% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombre d'adjoints}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{Un data.frame contenant les donnees des elus, incluant la colonne \code{Libelle_fonction}
qui specifie la fonction des elus.}
}
\value{
Un entier representant le nombre d'adjoints dans le data.frame.
}
\description{
Cette fonction permet de compter le nombre d'adjoints dans un jeu de donnees en fonction
de la colonne \code{Libelle_fonction}. Elle identifie les lignes contenant le terme "adjoint"
et retourne le nombre total d'adjoints dans le data.frame.
}
\examples{
df <- data.frame(Libelle_fonction = c("Maire", "Adjoint au Maire", "Adjoint", "Conseiller"))
compter_nombre_d_adjoints(df)

}
