% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{creer_departement}
\alias{creer_departement}
\title{Creer une Commune}
\usage{
creer_departement(df)

creer_departement(df)
}
\arguments{
\item{nom_commune}{Le nom de la commune à créer (caractères).}

\item{elus}{Un data.frame contenant les élus associés à la commune. Le data.frame doit comporter les colonnes
\code{Nom_elu}, \code{Prenom_elu}, \code{Libelle_fonction}, et \code{Date_de_naissance}.}

\item{nom_departement}{Le nom du département à créer (caractères).}

\item{code_departement}{Le code du département (caractères).}

\item{communes}{Une liste de data.frames représentant les communes du département. Chaque data.frame doit inclure les
informations des élus dans la commune.}
}
\value{
Un data.frame représentant la commune avec ses élus et les informations associées.

Un data.frame représentant le département avec ses communes et les élus associés.
}
\description{
Cette fonction permet de créer une commune en utilisant les informations passées en argument.
Elle crée un data.frame avec les colonnes nécessaires pour enregistrer les données relatives à la commune,
telles que son nom, son département, et les élus qui y figurent.

Cette fonction permet de créer un département en utilisant les informations fournies, telles que le nom du département,
le code du département et les communes qui en font partie.
}
\details{
La fonction vérifie d'abord que les données fournies pour les élus sont correctes et, si tout est en ordre,
elle crée une nouvelle commune avec les informations données. La commune inclura également une colonne avec
les informations des élus.

La fonction crée un département en prenant en compte les communes qui en font partie, et inclut un enregistrement
des élus pour chaque commune. Le département aura une colonne pour le nom, le code, ainsi que les communes associées.
}
\examples{
df_elus_commune <- data.frame(
  Nom_elu = c("Dupont", "Martin"),
  Prenom_elu = c("Jean", "Paul"),
  Libelle_fonction = c("Maire", "Conseiller"),
  Date_de_naissance = c("12/06/1960", "14/11/1970")
)
creer_commune("Nantes", "44", df_elus_commune)
df_elus_commune1 <- data.frame(
  Nom_elu = c("Dupont", "Martin"),
  Prenom_elu = c("Jean", "Paul"),
  Libelle_fonction = c("Maire", "Conseiller"),
  Date_de_naissance = c("12/06/1960", "14/11/1970")
)
df_elus_commune2 <- data.frame(
  Nom_elu = c("Bernard", "Lemoine"),
  Prenom_elu = c("Pierre", "Michel"),
  Libelle_fonction = c("Adjoint au Maire", "Conseiller"),
  Date_de_naissance = c("05/03/1980", "01/08/1985")
)
communes <- list(creer_commune("Nantes", "44", df_elus_commune1),
                 creer_commune("Faverelles", "44", df_elus_commune2))
creer_departement("Loire-Atlantique", "44", communes)
}
