% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{elu_sample}
\alias{elu_sample}
\title{Elus Sample Dataset}
\format{
\subsection{\code{elu_sample}}{

Un \code{data.frame} avec 1000 lignes et les colonnes suivantes :
\describe{
\item{Nom_elu}{Nom de l'elu}
\item{Prenom_elu}{Prenom de l'elu}
\item{Libelle_commune}{Nom de la commune de l'elu}
\item{Libelle_fonction}{Fonction de l'elu (par exemple, maire, adjoint)}
\item{Date_de_naissance}{Date de naissance de l'elu au format "JJ/MM/AAAA"}
\item{Autres colonnes}{Autres informations relatives à l'elu, telles que les informations de mandat, etc.}
}
}
}
\source{
Donnees generees pour l'exemple.
}
\usage{
elu_sample
}
\description{
Un jeu de donnees d'exemple contenant des informations sur les elus locaux.
}
\keyword{datasets}
