% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elus_conseillers_municipaux_cm}
\alias{elus_conseillers_municipaux_cm}
\title{Donnees des elus municipaux}
\format{
Un data frame avec 9753 lignes et 16 colonnes :
\describe{
\item{Code du departement}{Code numerique du departement}
\item{Libelle du departement}{Nom du departement}
\item{Code de la collectivite}{Code unique de la collectivite}
\item{...}{Autres colonnes}
}
}
\source{
\url{https://exemple.com/donnees}
}
\usage{
elus_conseillers_municipaux_cm
}
\description{
Ce jeu de donnees contient les informations sur les conseillers municipaux.
}
\keyword{datasets}
