% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Generer un rapport}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{La commune pour laquelle generer le rapport (code INSEE).}

\item{departement}{Le departement concerne (code numerique).}

\item{output}{Le nom du fichier de sortie (ex: "rapport.pdf" ou "rapport.html").}
}
\value{
Le chemin du fichier genere.
}
\description{
Cette fonction genere un rapport Quarto au format PDF ou HTML incluant :
}
