% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{plot.departement}
\alias{plot.departement}
\title{Visualiser la repartition des categories socio-professionnelles des Maires}
\usage{
\method{plot}{departement}(df)
}
\arguments{
\item{df}{Un data.frame contenant les données des élus, avec notamment les colonnes
\code{Libelle_fonction} (pour la fonction de l'élu), \code{Code_categorie_socio_professionnelle}
(pour les catégories socio-professionnelles), et \code{Libelle_commune} (pour les communes).}
}
\value{
Aucun retour. La fonction affiche un graphique en barres représentant les catégories
socio-professionnelles des Maires dans le département.
}
\description{
Cette fonction permet d'afficher un graphique en barres (barplot) représentant les 10
catégories socio-professionnelles les plus fréquentes parmi les Maires dans un département.
La fonction filtre les Maires dans le data.frame, exclut les adjoints et Maires délégués, puis
crée un barplot des catégories socio-professionnelles des Maires restants.
}
\examples{
df <- data.frame(
  Libelle_fonction = c("Maire", "Adjoint au Maire", "Maire", "Maire délégué"),
  Code_categorie_socio_professionnelle = c("CSP1", "CSP2", "CSP1", "CSP3"),
  Libelle_commune = c("Nantes", "Nantes", "Faverelles", "Faverelles"),
  Libelle_departement = c("Loire-Atlantique", "Loire-Atlantique", "Gers", "Gers")
)
plot.departement(df)
}
