% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Visualiser la répartition des catégories socio-professionnelles des maires}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un data.frame contenant les donnees des elus, incluant la colonne \code{Libelle_fonction}
qui specifie la fonction des elus et \code{Code_categorie_socio_professionnelle} pour les catégories
socio-professionnelles des élus.}
}
\value{
Un graphique en barres représentant la répartition des catégories socio-professionnelles des maires.
}
\description{
Cette fonction génère un graphique en barres horizontales montrant la répartition des catégories
socio-professionnelles des maires (excluant les adjoints et les délégués) dans un jeu de données.
Le graphique est basé sur la colonne \code{Code_categorie_socio_professionnelle} et les données sont
filtrées pour ne conserver que les maires, excluant les adjoints et les délégués.
}
\examples{
df <- data.frame(Libelle_fonction = c("Maire", "Adjoint au Maire", "Maire délégué", "Maire"),
                 Code_categorie_socio_professionnelle = c("Cadre", "Ouvrier", "Cadre", "Cadre"),
                 Code_departement = c("44", "44", "44", "44"),
                 Libelle_departement = c("Loire-Atlantique", "Loire-Atlantique", "Loire-Atlantique", "Loire-Atlantique"))
plot_code_professions(df)

}
