% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Trouver l'elu le plus age}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{Un data.frame contenant les donnees des elus, incluant la colonne \code{Date_de_naissance}
qui specifie la date de naissance des elus.}
}
\value{
Un data.frame avec le nom, prenom et la date de naissance de l'elu le plus age.
}
\description{
Cette fonction permet de trouver l'elu ayant l'age le plus avance dans un jeu de donnees.
Elle utilise la colonne \code{Date_de_naissance}, convertit les dates en format \code{Date}, puis
retourne le nom, le prenom et la date de naissance de l'elu le plus age.
}
\examples{
df <- data.frame(Nom_elu = c("Dupont", "Durand", "Martin"),
                 Prenom_elu = c("Jean", "Pierre", "Michel"),
                 Date_de_naissance = c("01/01/1980", "15/05/1975", "23/07/1990"))
trouver_l_elu_le_plus_age(df)

}
