% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{plot.commune}
\alias{plot.commune}
\title{Visualiser les catégories socio-professionnelles des Maires d'une commune}
\usage{
\method{plot}{commune}(df)
}
\arguments{
\item{df}{Un data.frame contenant les informations des élus d'une commune, avec au moins les colonnes
suivantes : \code{Libelle_fonction}, \code{Code_categorie_socio_professionnelle}, \code{Libelle_commune}, et \code{Libelle_departement}.
La colonne \code{Libelle_fonction} doit inclure les titres des fonctions des élus (Maire, adjoint, etc.).}
}
\value{
Un graphique de type barplot, représentant la fréquence d'apparition des différentes catégories socio-professionnelles
des Maires.
}
\description{
Cette fonction génère un barplot représentant la fréquence des différentes catégories socio-professionnelles
des Maires d'une commune, en filtrant les Maires et en excluant les adjoints et les délégués.
Le barplot est affiché avec des barres horizontales et inclut les codes des catégories socio-professionnelles.
}
\examples{
df <- data.frame(Libelle_fonction = c("Maire", "Maire-adjoint", "Conseiller", "Maire"),
                 Code_categorie_socio_professionnelle = c("Cadre", "Ouvrier", "Cadre", "Cadre"),
                 Libelle_commune = "Nantes",
                 Libelle_departement = "Loire-Atlantique")
plot.commune(df)
}
